code   把详情页点赞的用户头像放到第一个

+++ app/src/main/java/com/zhiyicx/thinksnsplus/modules/dynamic/detail/DynamicDetailPresenter.java
@@ -374,8 +374,7 @@ public class DynamicDetailPresenter extends AppBasePresenter<DynamicDetailContra
                 }
             }
         } else {// 喜欢
-            UserInfoBean mineUserInfo = mUserInfoBeanGreenDao.getSingleDataFromCache((long)
-                    AppApplication.getMyUserIdWithdefault());
+            UserInfoBean mineUserInfo = mUserInfoBeanGreenDao.getSingleDataFromCache(AppApplication.getMyUserIdWithdefault());
             DynamicDigListBean dynamicDigListBean = new DynamicDigListBean();
             dynamicDigListBean.setUser_id(mineUserInfo.getUser_id());
             dynamicDigListBean.setId(System.currentTimeMillis());
